+++ src/experimental/index.ts
@@ -1,2 +1,4 @@
 export * from "./preference.js";
 export * from "./advanced-distillation.js";
+export * from "./scale.js";
+export * from "./governed-logs.js";